- Change ref to reactive - Made changes in route for loginpage - Removed unnecesarry code - Rename SigninPage.vue file to loginPage.vue
